refactor: rename alertErrorWait to alertWaitOpaque and simplify implementation
- Rename alertErrorWait to alertWaitOpaque for clarity (reflects opaque background variant)
- Remove async/await from alertWaitOpaque function (no user interaction needed)
- Remove await when calling alertWaitOpaque (display only, no blocking)
- Rename drive.ts to drive.svelte.ts for consistency
- Update all import paths from drive/drive to drive/drive.svelte

The function displays information with an opaque background for Tauri OAuth flow,
where users copy the auth code and return to the Tauri app. No user confirmation
is needed in the web context.
