+++ src/lib/Setting/Pages/FilesSettings.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import { language } from "src/lang";
     import { alertConfirm } from "src/ts/alert.svelte";
-    import { checkDriver } from "src/ts/drive/drive";
+    import { checkDriver } from "src/ts/drive/drive.svelte";
     import { isTauri, isNodeServer } from "src/ts/platform"
 
 </script>
