+++ src/lib/Setting/Pages/UserSettings.svelte
@@ -9,7 +9,7 @@
     import { forageStorage, loadInternalBackup } from "src/ts/globalApi.svelte";
     import { isTauri, isNodeServer, isCapacitor } from "src/ts/platform"
     import { unMigrationAccount } from "src/ts/storage/accountStorage";
-    import { checkDriver } from "src/ts/drive/drive";
+    import { checkDriver } from "src/ts/drive/drive.svelte";
     import { LoadLocalBackup, SaveLocalBackup } from "src/ts/drive/backuplocal";
     import Button from "src/lib/UI/GUI/Button.svelte";
     import { exportAsDataset } from "src/ts/storage/exportAsDataset";
